fix(views): every list shares a wide terminal's leftover width
The previous commit made the trailing grow column absorb the leftover only
when it had something to spend it on. That fixed services, configs, secrets
and nodes, and left volumes and charts exactly as they were, because a mount
path and a convergence detail are never empty — volumes on a 360-column
terminal still gave MOUNT POINT a 300-cell column to hold a 43-character
path and packed the rest into the left third.

So the rule goes rather than the opt-in: Column.Grow is removed and Flex
carries both meanings again, as it did before #575. Every content-aware view
— services, configs, secrets, nodes, volumes and both charts tables — shares
the surplus across its elastic columns at every width.

This reopens what #575 set out to fix: a 13-character volume name sits in a
146-cell column at 360 columns. That is the cost of filling the width, and
filling it is what operators asked for.

charts keeps its withDetail flex gate. That gate is about shrinking, not
filling: minDetailWidth is 28 while a detail string is typically ~30, so the
shrink path runs routinely, and making NAME and CHART elastic there would
win the explanation its width by truncating the release's identity.
