+++ ui/components/filterable/list/layout_test.go
@@ -21,6 +21,16 @@ func testCols() []Column[row] {
 	}
 }
 
+// oneFlexCols is a second shape for the width invariants: a single elastic
+// column, so a rule that only holds when the slack divides evenly is caught.
+func oneFlexCols() []Column[row] {
+	return []Column[row]{
+		{Label: "NAME", MinWidth: 4, Flex: true, Cell: func(r row) string { return r.a }},
+		{Label: "KIND", MinWidth: 4, Cell: func(r row) string { return r.b }},
+		{Label: "LABELS", MinWidth: 6, Cell: func(r row) string { return r.c }},
+	}
+}
+
 func sum(xs []int) int {
 	t := 0
 	for _, x := range xs {
@@ -138,50 +148,11 @@ func TestScrollRightStopsAtTheEndOfTheText(t *testing.T) {
 		"so the tail is on screen rather than scrolled past")
 }
 
-// growCols separates the two elasticities: NAME gives up width when squeezed
-// and scrolls when truncated, while the trailing column takes the leftover.
-func growCols() []Column[row] {
-	return []Column[row]{
-		{Label: "NAME", MinWidth: 4, Flex: true, Cell: func(r row) string { return r.a }},
-		{Label: "KIND", MinWidth: 4, Cell: func(r row) string { return r.b }},
-		{Label: "LABELS", MinWidth: 6, Grow: true, Cell: func(r row) string { return r.c }},
-	}
-}
-
-// Slack goes to the Grow column, so the columns before it stay put however wide
-// the terminal gets. A table of short values is where that matters: growing a
-// middle column opens a void in the middle of every row.
-func TestLayoutWidths_GrowAbsorbsSlackNotFlex(t *testing.T) {
-	items := []row{{a: "short", b: "kind", c: "team=platform"}}
-
-	narrow := LayoutWidths(growCols(), items, 60, -1)
-	wide := LayoutWidths(growCols(), items, 200, -1)
-
-	require.Equal(t, narrow[0], wide[0], "the flex column must not absorb slack")
-	require.Equal(t, narrow[1], wide[1])
-	require.Greater(t, wide[2], narrow[2], "the grow column takes it instead")
-	require.Equal(t, 200, sum(wide), "and the row still spans the width")
-}
-
-// A grow column every one of whose cells fits inside its own header has nothing
-// to spend the leftover on, so putting it there would leave the terminal's right
-// half dead rather than fill it — services' ERROR on a swarm with no errors is
-// the case. The Flex columns take it back.
-func TestLayoutWidths_EmptyGrowColumnHandsSlackBackToFlex(t *testing.T) {
-	items := []row{{a: "short", b: "kind", c: ""}}
-
-	narrow := LayoutWidths(growCols(), items, 60, -1)
-	wide := LayoutWidths(growCols(), items, 200, -1)
-
-	require.Greater(t, wide[0], narrow[0], "the flex column absorbs instead")
-	require.Equal(t, narrow[1], wide[1], "the non-elastic column stays put")
-	require.Equal(t, narrow[2], wide[2], "and the empty grow column takes none of it")
-	require.Equal(t, 200, sum(wide), "the row still spans the width")
-}
-
-// A view that declares no Grow keeps the behaviour it had before Grow existed,
-// so adding the field changed nothing for the views already using Flex.
-func TestLayoutWidths_FlexStillAbsorbsWhenNoGrowDeclared(t *testing.T) {
+// The Flex columns share a wide terminal's leftover between them, so the table
+// spans the width instead of packing into the left of it. Confining the whole
+// surplus to the trailing column was tried instead and reverted: it stops the
+// columns before it drifting apart by leaving the rest of the screen dead.
+func TestLayoutWidths_FlexColumnsShareTheLeftover(t *testing.T) {
 	items := []row{{a: "short", b: "kind", c: "x=1"}}
 
 	narrow := LayoutWidths(testCols(), items, 60, -1)
@@ -192,24 +163,12 @@ func TestLayoutWidths_FlexStillAbsorbsWhenNoGrowDeclared(t *testing.T) {
 	require.Equal(t, 200, sum(wide))
 }
 
-// Grow is about width the table does not need; it must not rescue a column from
-// shrinking when the terminal is too narrow for the content.
-func TestLayoutWidths_GrowColumnStillShrinksWhenNarrow(t *testing.T) {
-	items := []row{{a: "a-fairly-long-name-here", b: "kind", c: "team=platform,env=prod"}}
-
-	w := LayoutWidths(growCols(), items, 40, -1)
-	require.LessOrEqual(t, sum(w), 40)
-	for i, c := range growCols() {
-		require.GreaterOrEqual(t, w[i]-ColGap, displayWidth(c.Label), "column %d must still fit its label", i)
-	}
-}
-
 // NaturalWidth must be the width at which LayoutWidths neither stretches nor
 // squeezes, or a view asking "have I room to spare?" gets a wrong answer.
 func TestNaturalWidth_IsTheNoStretchNoSqueezeWidth(t *testing.T) {
 	items := []row{{a: "some-name", b: "kind", c: "team=platform"}}
 
-	for _, cols := range [][]Column[row]{testCols(), growCols()} {
+	for _, cols := range [][]Column[row]{testCols(), oneFlexCols()} {
 		natural := NaturalWidth(cols, items, -1)
 		require.Equal(t, natural, sum(LayoutWidths(cols, items, natural, -1)),
 			"at the natural width the layout is an identity")
